# DNN: Further optimization of Conv2D