--- a/modules/dnn/src/layers/fast_convolution/fast_convolution.hpp
+++ b/modules/dnn/src/layers/fast_convolution/fast_convolution.hpp
@@ -7,20 +7,26 @@
 
 #include "opencv2/core/hal/intrin.hpp"
 
-#ifndef FAST_CONV_PRAM
-#define FAST_CONV_PRAM
+#ifndef CONV_PRAM
+#define CONV_PRAM
 #if CV_NEON && CV_NEON_AARCH64  // 32 registers.
-#define FAST_CONV_MR 4
-#define FAST_CONV_NR 28
+#define CONV_MR 4
+#define CONV_NR 28
 enum { FAST_VEC_NLANES=4 };
 #elif CV_NEON              // 16 registers.
-#define FAST_CONV_MR 4
-#define FAST_CONV_NR 12
+#define CONV_MR 4
+#define CONV_NR 12
 enum { FAST_VEC_NLANES=4 };
 #else // SIMD 128, AVX or AVX2
-#define FAST_CONV_MR 4
-#define FAST_CONV_NR 24
-enum { FAST_VEC_NLANES=4 };
+#define CONV_MR 4
+#define CONV_NR 24
+
+#ifdef CV_AVX2
+enum { FAST_VEC_NLANES=8 }; // AVX2
+#else
+enum { FAST_VEC_NLANES=4 }; // SIMD 128
+#endif
+
 #endif
 #endif
 
@@ -37,7 +43,6 @@ struct FastConv2d
 
     std::vector<float> weightsBuf;        // For generic Conv 2D
     std::vector<float> weightsWino63Buf;  // For Winograd F(6x6, 3x3).
-
     std::vector<float> biasBuf;
     bool ifWinograd63 = false;
     bool useAVX2 = checkHardwareSupport(CPU_AVX2);
@@ -52,20 +57,20 @@ Ptr<FastConv2d> initFastConv2d(
         int dilation_x, int dilation_y,
         const std::vector<size_t>& pads_begin,
         const std::vector<size_t>& pads_end,
-        float* srcWeights,
+        InputArray weightsMat,
         float* srcBias);
 
 // It contains different computing branches, like winograd, 1x1 conv.
-void runFastConv2d(InputArray _input, OutputArray _output,
-                 const Ptr<FastConv2d>& conv, int ntasks, const Ptr<ActivationLayer>& actLayer);
+void runFastConv2d(InputArray _input, OutputArray _output, const Ptr<FastConv2d>& conv, int ntasks,
+                   const Ptr<ActivationLayer>& actLayer, bool fusedAdd);
 
 void runDepthwise(InputArray _input, OutputArray _output, const Ptr<FastConv2d>& conv, float minval, float maxval,
         ActivationLayer* activ, bool ifMinMaxAct);
 
 // winograd init
-void initWinograd63(Ptr<FastConv2d>& conv, float* src_weight, int K, int C);
+void initWinograd63(Ptr<FastConv2d>& conv, InputArray weightsMat, int K, int C);
 
-int runWinograd63(InputArray _input, OutputArray _output, const Ptr<FastConv2d>& conv, int ntasks,
+int runWinograd63(InputArray _input, InputArray _fusedAddMat, OutputArray _output, const Ptr<FastConv2d>& conv, int ntasks,
                   float minval, float maxval, ActivationLayer* activ, bool ifMinMaxAct);
 
 } // namespace dnn
@@ -73,9 +78,7 @@ int runWinograd63(InputArray _input, OutputArray _output, const Ptr<FastConv2d>&
 namespace opt_AVX2
 {
 #if CV_TRY_AVX2
-void convBlock_AVX2(int k, const float *a, const float *b,
-        float *c, int ldc, const float *bias,
-        float minval, float maxval, bool ifActiv);
+void convBlock_AVX2(int np, const float* a, const float* b, float* c, int ldc, bool init_c);
 
 void depthWiseBlock_AVX2(const float *inptr, float *outptr, const float *weights, float biasval, int *ofstab, int *yxtab,
                 float minval, float maxval, int Hi, int Wi, int H0, int W0, int ksize, int pad_top, int pad_left,