@asenyaev 2022-08-29
Seems that merging this PR a build with coverage started to fail ([log](https://github.com/opencv/ci-gha-workflow/runs/8065443688?check_suite_focus=true#step:11:899)).